requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cmake==3.31.4
colorama==0.4.6
cryptography==44.0.1
dlib==19.24.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.8
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jwt==1.3.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.0.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.0
starlette==0.45.3
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
